chore(pkgs/bench): move `bench` to `packages`
