+++ package.json
@@ -7,8 +7,7 @@
   "workspaces": [
     "crates/rolldown_binding",
     "packages/**",
-    "web/**",
-    "bench"
+    "web/**"
   ],
   "scripts": {
     "lint": "oxlint --ignore-path=./.oxlintignore --import-plugin",
